+++ src/Core/src/EnvVar.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AsyncAws\Core;
+
+final class EnvVar
+{
+    public static function get(string $name): ?string
+    {
+        if (isset($_ENV[$name])) {
+            // variable_order = *E*GPCS
+            return $_ENV[$name];
+        } elseif (isset($_SERVER[$name]) && 0 !== strpos($name, 'HTTP_')) {
+            // fastcgi_param, env var, ...
+            return $_SERVER[$name];
+        } elseif (false === ($env = getenv($name)) || null === $env) {
+            // getenv not thread safe
+            return null;
+        }
+
+        return $env;
+    }
+}